test: update Tests affected from implementing ChatWebSocketController
